devworld/requirements.txt
django-environ==0.9.0
djangorestframework
djangorestframework-simplejwt
Django
pillow
gunicorn
environ
django-bootstrap-v5
django-environ
fontawesomefree
psycopg2-binary
